Write q_insert_head() and q_insert_tail()
The contents in these two functions are nearly identical.
The allocated factor is size+1 since strlen() does not count the null
terminator. Otherwise, sizeof(char) should not be neglected since it may
not be 1.
We use strncpy() because it is safer than strcpy().
The list operations are put at last because possible failures first
would make the return-false conditions easier.