Write q_delete_duplicate()
Introduce another list `trashcan` to list all nodes to delete, and
delete them at last. In this way, we do not have to call free() during
traversing.

The first entry of each group of duplicates should not be delete too
soon, otherwise accessing curr would cause error.

Maybe we need a better way to prevent the return due to trashcan.