Write q_reverse() and add swap_pointer()
q_reverse():
Since the list is doubly-linked, we may reverse it by swapping the
pointers of every node. The loop should not start from head, otherwise,
it would end at first.

swapptr():
Use char type in order to be compatible with pointer to any type.
I tried to swap by using xor operation, but I got an error. Need to
figure it out.